requirements.txt
dlib==19.24.6
focal_loss==0.0.7
graphviz==0.20.3
imutils==0.5.4
librosa==0.10.2.post1
lpips==0.1.4
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.10.0.84
Pillow==11.1.0
pydub==0.25.1
scikit_learn==1.6.1
scipy==1.15.2
setuptools==75.1.0
skimage==0.0
tensorflow==2.18.0
tf_slim==1.1.0
torch==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.67.1
visdom==0.2.4
wavfile==4.7.2
